Split persist read/write into separate files for easier testing.
